refactor(peagen): split monolithic core
