+++ pkgs/standards/peagen/peagen/core/__init__.py
@@ -1,0 +1,17 @@
+"""Composable Peagen engine."""
+from colorama import Fore, Style
+
+from .._config import _config
+from .._graph import _topological_sort, _transitive_dependency_sort
+from .._utils._processing import _process_project_files
+from .engine import Peagen
+
+__all__ = [
+    "Peagen",
+    "Fore",
+    "Style",
+    "_config",
+    "_topological_sort",
+    "_transitive_dependency_sort",
+    "_process_project_files",
+]
